fix waybill export-import
