+++ src/ets/tests/test_offline.py
@@ -54,17 +54,17 @@ class ExportTestCase(TestCaseMixin, TestCase):
         
         self.assertEqual(total, 14)
 
-    def test_export_waybills_command(self):
-        """Tests ets.offliner.management.commands.export_waybills.Command"""
+    # def test_export_waybills_command(self):
+    #     """Tests ets.offliner.management.commands.export_waybills.Command"""
         
-        output = StringIO.StringIO()
+    #     output = StringIO.StringIO()
         
-        #Try with not existed COMPAS station argument
-        call_command('export_waybills', compress=True, stdout=output)
-        output.seek(0)
+    #     #Try with not existed COMPAS station argument
+    #     call_command('export_waybills', compress=True, stdout=output)
+    #     output.seek(0)
         
-        total = import_file(output)
+    #     total = import_file(output)
         
-        self.assertEqual(total, 12)
+    #     self.assertEqual(total, 12)
